requirements.txt
torch
streamlit
torchvision
streamlit
Pillow
